+++ lib/gecko/record/note.rb
@@ -1,0 +1,15 @@
+require 'gecko/record/base'
+
+module Gecko
+  module Record
+    class Note < Base
+      belongs_to :company
+      belongs_to :user
+
+      attribute :body,  String
+    end
+
+    class NoteAdapter < BaseAdapter
+    end
+  end
+end