+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Add `Gecko::Record::Note` model.
+
 ## 0.7.0 (2019-02-19)
 - Add new `country_code` attribute to Address model.
 - Stop testing on Ruby 2.1 and 2.2